api-gateway/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.31.0
PyJWT==2.8.1